Merge overview branch + fixes

+++ grails-app/controllers/au/org/ala/volunteer/ValidationRuleController.groovy
@@ -5,16 +5,33 @@ import au.org.ala.web.AlaSecured
 @AlaSecured(value = ["ROLE_VP_ADMIN"], redirectController = "index")
 class ValidationRuleController {
 
+    def userService
+
     def index() {
         redirect(action:list())
     }
 
+    boolean checkAdmin() {
+        if(!userService.isAdmin()) {
+            flash.message = "You do not have permission to view this page"
+            redirect(url: "/")
+            return false
+        }
+        return true
+    }
+
     def list() {
+        if (!checkAdmin()) {
+            return
+        }
         def validationRules = ValidationRule.list(params)
         [validationRules: validationRules, totalCount: ValidationRule.count ]
     }
 
     def delete() {
+        if (!checkAdmin()) {
+            return
+        }
         def rule = ValidationRule.get(params.int("id"));
         if (rule) {
             rule.delete()
@@ -26,16 +43,25 @@ class ValidationRuleController {
     }
 
     def addRule() {
+        if (!checkAdmin()) {
+            return
+        }
         def rule = new ValidationRule(name:"<New rule>")
         render(view: 'edit', model: [rule: rule])
     }
 
     def edit() {
+        if (!checkAdmin()) {
+            return
+        }
         def rule = ValidationRule.get(params.int("id"))
         [rule: rule]
     }
 
     def update() {
+        if (!checkAdmin()) {
+            return
+        }
         def rule = ValidationRule.get(params.int("id"))
         if (rule) {
             rule.properties = params
